requirements.txt
2captcha-python==1.2.2
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.8.2
Brotli==1.1.0
captcha==0.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Deprecated==1.2.14
discloud==1.0.1
discord-buttons-plugin==0.1.2
discord.js==1.0.7
discord.py==2.3.2
Django==4.2.6
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.3
frozenlist==1.4.0
future==1.0.0
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.2.0
javascriptpy==1.1.6
Jinja2==3.1.4
keep==2.10.1
keep-alive==0.0.1
MarkupSafe==2.1.5
multidict==6.0.4
mutagen==1.47.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
payload-api==0.4.10
Pillow==10.1.0
pycparser==2.21
pycryptodomex==3.20.0
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
pytchat==0.5.5
python-dateutil==2.8.2
python-hosts==1.0.4
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
South==1.0.2
sqlparse==0.4.4
terminaltables==3.1.10
tweepy==4.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Voice==0.1.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.15.0
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2024.4.9